moving marbles
Create an animation on the page that makes an orange marble move to the next location in the line every second. Allow the user to stop the animation by placing the cursor on any marble. The animation will restart again once the user remove the cursor from that marble. Add your own interesting feature to the script that tinkers with the speed or location of images